requirements.txt
requests==2.28.2
beautifulsoup4=4.12.0
lxml==4.9.2
pydantic==1.10.7
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pillow==9.5.0
celery==5.2.7
django-celery-beat==2.5.0
django-redis==5.2.0
django-celery-results==2.5.0